refactor: split arcane table into multiple files
