+++ frontend/src/lib/components/arcane-table/arcane-table-cell.svelte
@@ -1,0 +1,5 @@
+<script lang="ts">
+	let { value }: { value: unknown } = $props();
+</script>
+
+<span class="max-w-[500px] truncate">{value ?? ''}</span>
